+++ examples/no_ecs.rs
@@ -51,12 +51,12 @@ fn main() {
     #[cfg(feature = "rust-alloc")]
     lv_bevy_ecs::malloc::set_mem_monitor(get_memory_stats);
 
-    const HOR_RES: u32 = 320;
-    const VER_RES: u32 = 240;
-    const LINE_HEIGHT: u32 = 16;
+    const HOR_RES: usize = 320;
+    const VER_RES: usize = 240;
+    const LINE_HEIGHT: usize = 16;
 
     let mut sim_display: SimulatorDisplay<Rgb565> =
-        SimulatorDisplay::new(Size::new(HOR_RES, VER_RES));
+        SimulatorDisplay::new(Size::new(HOR_RES as u32, VER_RES as u32));
 
     let output_settings = OutputSettingsBuilder::new().scale(1).build();
     let mut window = Window::new("Button Example", &output_settings);
